use new 'serviceNames' configuration object
Seems like we're passing 'serviceName' to nunjucks in a lot of places
for no reason - we already use `getFullServiceName()` to map a
short service name (called 'serviceType' in the templates) to a full
service name. Switching to a map of service names allows us to skip
passing the service name explicitly.