content changes related to dataset 'status'
- arbitrary strings that were sometimes used, sometimes used
in combination with other columns, even when it looked like the same
information was needed (e.g. `status`, which had values inconsistent
with the ones we actually want to surface to the users, I can't find
any reason why we would want them to be different)
- some post-processing of the data was done (e.g. adding `error`
and `issue` columns in code) - I removed that and now the query returns
the `error` column, the issue column was removed.
- remove needless checks from `perfomanceDbApi.test.js` - another issue
'testing mock data' - no real benefit of doing extensive tests here,
better taken care of in other tests (e.g. in `lpaOverviewPage.test.js`)