Remove view "dataset" link
This was put in place as a stopgap when the service didn't have the capabilities to play back data to LPAs.

This should now be removed, as it creates a disjointed journey for users. Usability testing sessions have evidenced the fact that people get lost when they click on the link.