gathering datasette query information
Adds a feature (that has to be enabled via config) to gather information
about datasette queries fired by the app.

If enabled, the information will be present under '/request-info' as
JSON.

To see how it works. enable the feature, start the app and click go to
a few pages, then see the output of '/request-info'

To make the above possible the the following changes were needed:
- add unique ID to every request
- update datasette.runQuery() to accept a third (optional) options
  param which can hold the request object. This allows to link the
  request and associated datasette queries
- update all the places where runQuery() is called
- also, the execution of remote datasette queries is timed and available
  in the JSON data returned from '/request-info'
- the metrics are not logged by default (it can be quite noisy), but
  adding a 'metric' query param to a URL will log them