Numbers will now be shown
The platform uses a lot of numbers which are sometimes relevant to the session. As the data is public this can be recorded.

The email and IP address is not shown intentionally as we do not want the data to be recorded. We have the submit form which records the email address.

Further information can be found here:
https://web.developer.smartlook.com/docs/consent-and-sensitive-data#combining-calls